# Conflicts:
#	hedera-node/hapi-utils/src/main/java/com/hedera/node/app/hapi/utils/ethereum/EthTxData.java
#	hedera-node/hapi-utils/src/main/java/com/hedera/node/app/hapi/utils/ethereum/EthTxSigs.java
#	hedera-node/hapi-utils/src/test/java/com/hedera/node/app/hapi/utils/ethereum/EthTxDataTest.java
#	hedera-node/hapi-utils/src/test/java/com/hedera/node/app/hapi/utils/ethereum/EthTxSigsTest.java
#	hedera-node/test-clients/src/main/java/com/hedera/services/bdd/spec/transactions/contract/HapiEthereumCall.java
#	hedera-node/test-clients/src/main/java/com/hedera/services/bdd/spec/transactions/contract/HapiEthereumContractCreate.java
#	hedera-node/test-clients/src/main/java/com/hedera/services/bdd/suites/contract/hapi/ContractCreateSuite.java
#	hedera-node/test-clients/src/main/java/com/hedera/services/bdd/utils/Signing.java
#	hedera-node/test-clients/src/test/java/com/hedera/services/bdd/utils/test/SigningTest.java

+++ hedera-node/test-clients/src/main/resources/validation-scenarios/Multipurpose.sol
@@ -1,3 +1,4 @@
+// SPDX-License-Identifier: Apache-2.0
 pragma solidity ^0.5.3;
 
 contract Multipurpose {
@@ -16,7 +17,7 @@ contract Multipurpose {
     return luckyNumber;
   }
 
-  function donate(uint32 toNum, string memory saying) public payable {
+  function donate(uint120 toNum, string memory saying) public payable {
     address payable beneficiary = address(uint120(toNum));
     beneficiary.transfer(1);
     emit Boast(saying);
